test(chaos): consolidate supervisor tests into dom selector chaos suite
